chore(deps): update eslint/prettier config
