+++ src/events/mineflayer/handler/error.ts
@@ -3,9 +3,8 @@ import winston from 'winston';
 export default {
     name: 'error',
     runOnce: false,
-    run: (bot, error: Error) => {
-        winston.error('Encountered an unexpected error. Restarting the bot in 15 seconds...');
-        winston.error(error);
+    run: (_bot, error: Error) => {
+        winston.error('Encountered an unexpected error. Exiting in 15 seconds...', error);
 
         setTimeout(() => {
             process.exit(1);
